WIP: Laying down the framework
I realized that we need a lot more than just funciton calls.
This commit helps start us off with mutliprocessing for the various
worker processess that we need. It also has a rudimentary tester that
automates the tedious command passing on the stdin. Will come in handy later.

This is WIP, the process spawn fine and run as desired. There's no functionality
for now but it lays down the frameworks as the title says.

Let's build on top of this. Pushing to master.

Signed-off-by: Simarpreet Singh <s244sing@uwaterloo.ca>