simulator: Improving collisionDetection with locking
So this commit makes each thread have exclusive access to the link.
At any given time, the time to transmit for a thread may be right, but
as long as it doesn't have the mutex it won't be able to and would therefore
qualify as a collision..

This isn't really the best approach as in the ideal case we can have two nodes
use the same link as long as their packets are not colliding with each other
on the physical queue (link). That's why we have the is_detected() logic for.

This is definitely a step forward than the previous collisionDetection() logic
but a bit too aggressive. If nothing good comes up, we should go with this.

Signed-off-by: Simarpreet Singh <simar@linux.com>